refactor(web): decompose ModelDiscrepancyModal into discrepancy/ components
ModelDiscrepancyModal was one 1180-line component. The render helpers it
closed over become components under components/discrepancy/, each with
its doc comment, and the modal keeps the state machine, the derived
values, the header and the confirm dialogs:

- groups.ts: the Group type, ALL_GROUPS and actionableIn, shared by the
  modal and the sections
- ClaimRow.tsx: one claim line with its flap chip, meta and tooltips;
  ClaimRowActions carries the modal-wide gating so rows cannot disagree
- BucketSection.tsx: the level-2 collapsible bucket line
- ClearedSummary.tsx: the dismissed/resolved headline over a cleared
  provider's buckets
- ProviderSection.tsx: the level-1 pill row plus its buckets, reading
  the modal's single open path
- GroupClaimsSection.tsx: the failover-group claims
- InformationalJournal.tsx: zone 2 with its entry renderer; expansion
  stays the modal's decision and comes in as props

No behaviour change: the data-testids, the single-open-path rule, the
mount-only-while-open buckets and the return-to-top observer are as
before, and the existing suites pass unchanged. The component drops
from 1180 to ~450 lines, under the max-lines-per-function cap, so its
eslint-disable goes; the file leaves the size allowlist (1276 -> 619).

Verified on the rebuilt dev stack: pills, bucket unroll, rows, journal
toggle and summary all render and behave as before.
