+++ scripts/ci/size-allowlist.txt
@@ -27,7 +27,6 @@ internal/proxy/proxy.go	976
 internal/proxy/upstream_classify.go	804
 web/src/api/types.ts	1058
 web/src/components/Layout.tsx	1065
-web/src/components/ModelDiscrepancyModal.tsx	1276
 web/src/components/VirtualModelTable.tsx	858
 web/src/components/__tests__/Layout.navigation.test.tsx	2558
 web/src/pages/Chat/useChat.ts	846
